requirements.txt
absl-py==1.1.0
altgraph==0.17.2
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.5
cryptography==37.0.2
cycler==0.11.0
DateTime==4.5
decorator==4.4.2
flatbuffers==1.12
fonttools==4.33.3
future==0.18.2
gast==0.4.0
gmail==0.6.3
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
google-pasta==0.2.0
googleapis-common-protos==1.56.4
grpcio==1.46.3
h11==0.13.0
h5py==3.7.0
httplib2==0.20.4
idna==3.3
imageio==2.21.0
imageio-ffmpeg==0.4.7
keras==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.4.3
libclang==14.0.1
llvmlite==0.38.1
loguru==0.5.3
Markdown==3.3.7
MouseInfo==0.1.3
moviepy==1.0.3
notify-py==0.3.3
notifypy==1.0.3.0
numba==0.55.2
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pause==0.3
pefile==2022.5.30
Pillow==9.2.0
proglog==0.1.10
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
PyJWT==2.4.0
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
pytube==12.1.0
pytweening==1.0.4
pytz==2022.1
pywin32==304
pywin32-ctypes==0.2.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.9.0
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sounddevice==0.4.4
SoundFile==0.10.3.post1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
twilio==7.12.0
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
win32-setctime==1.1.0
wrapt==1.14.1
wsproto==1.1.0
zope.interface==5.4.0
